[Refactor] #25 - 로그인 액티비티 및 메인 액티비티 연결
