+++ app/build.gradle
@@ -3,6 +3,7 @@ plugins {
     id ("kotlin-android")
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
+    id("com.google.devtools.ksp") version("1.8.0-1.0.9")
 }
 
 android {
@@ -72,10 +73,12 @@ dependencies {
 
     //room
     implementation "androidx.room:room-ktx:$room_version"
-    implementation "androidx.room:room-runtime:$room_version"
-    kapt "androidx.room:room-compiler:$room_version"
+    implementation("androidx.room:room-runtime:$room_version")
+    annotationProcessor("androidx.room:room-compiler:$room_version")
+    kapt("androidx.room:room-compiler:$room_version")
     implementation "androidx.room:room-rxjava2:$room_version"
     testImplementation "androidx.room:room-testing:$room_version"
+    ksp("androidx.room:room-compiler:$room_version")
 
     // viewpager2
     implementation "androidx.viewpager2:viewpager2:$viewpager2_version"
