+++ app/src/main/java/org/cazait/cazait_android/data/model/local/AppDatabase.kt
@@ -3,7 +3,7 @@ package org.cazait.cazait_android.data.model.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 
-@Database(entities = [Log::class], version = 1, exportSchema = false)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun logDao(): LogDao
-}
+//@Database(entities = [Log::class], version = 1, exportSchema = false)
+//abstract class AppDatabase : RoomDatabase() {
+//    abstract fun logDao(): LogDao
+//}
