+++ build.gradle
@@ -44,7 +44,7 @@ buildscript {
 plugins {
     id("com.android.application") version '7.4.0' apply false
     id("com.android.library") version '7.4.0' apply false
-    id("org.jetbrains.kotlin.android") version '1.7.20' apply false
+    id("org.jetbrains.kotlin.android") version '1.8.0' apply false
     id("com.google.dagger.hilt.android") version "2.44" apply false
 
 }
